requirements.txt
python-dotenv
python-multipart
langchain
langchain-community
langchain-experimental
langchain-openai
langchain-ollama
langchain-groq
colorama
langserve
fastapi
uvicorn
starlette
sqlalchemy
tortoise-orm
tortoise
asyncpg
web3
pinata
requests-toolbelt
PyPDF2
pypdf